cleaned up codes
